requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
bert_score==0.3.13
datasets==2.18.0
evaluate==0.4.2
fschat==0.2.36
huggingface_hub==0.23.2
huggingface_hub==0.21.3
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.32.3
rewardbench==0.1.1
scipy==1.13.1
seaborn==0.13.2
sentence_transformers==3.0.0
setuptools==68.2.2
torch==2.2.1
tqdm==4.66.2
tqdm==4.66.4
transformers==4.40.0
transformers==4.41.2
vllm==0.4.3
